# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go 语言编程之旅:一起用Go写项目",
        "title": "web-gin",
        "termsOfService": "https://github.com/spider-nns/web-gin",
        "contact": {},
        "license": {},
        "version": "v1"
    },
    "paths": {
        "/api/v1/tags": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "查询标签列表",
                "parameters": [
                    {
                        "description": "标签",
                        "name": "name",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "状态",
                        "name": "state",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "页，默认1",
                        "name": "pageNo",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "页大小，默认10",
                        "name": "pageSize",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ]
            }
        },
        "/api/v1/tags/": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "创建标签",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/app.SOut"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/err.Resp"
                        }
                    }
                }
            }
        },
        "/api/v1/tags/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "查询标签",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/app.SOut"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/err.Resp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.SOut": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "err.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}